Added a .sql file.
Added a .sql file that contains the basic creation of table. Further updates can change the inner code. Currently three tables are there in it. Account, Customer and Pan in the structured. The code is written in such a way that it should be run in up to down series.